+++ src/test/java/edu/brown/cs/fork/database/QueryUsersTest.java
@@ -32,19 +32,19 @@ public class QueryUsersTest implements ITest {
     setUp();
 
     try {
-      boolean success = this.db.updateUserGottenWay("ed", 0.2f);
+      boolean success = this.db.updateUserGottenWay("ed", 0.2);
       assertTrue(success);
-      float gottenWay = this.db.getUserGottenWay("ed");
+      double gottenWay = this.db.getUserGottenWay("ed");
       assertEquals(gottenWay, 0.2, 0.00001);
-      success = this.db.updateUserGottenWay("ed", 0.3f);
+      success = this.db.updateUserGottenWay("ed", 0.3);
       assertTrue(success);
       gottenWay = this.db.getUserGottenWay("ed");
       assertEquals(gottenWay, 0.3, 0.00001);
-      success = this.db.updateUserGottenWay("alan", 0.65f);
+      success = this.db.updateUserGottenWay("alan", 0.65);
       assertTrue(success);
       gottenWay = this.db.getUserGottenWay("alan");
       assertEquals(gottenWay, 0.65, 0.00001);
-      success = this.db.updateUserGottenWay("alan", 0.11f);
+      success = this.db.updateUserGottenWay("alan", 0.11);
       assertTrue(success);
       gottenWay = this.db.getUserGottenWay("alan");
       assertEquals(gottenWay, 0.11, 0.00001);
